Home.vue -> FileBrowser.vue, add DandisetList
